Remove the shitty time method
The post will grab the date time from the file name instead of the date attribute in the front-matter. When the "post" still a draft the file will not contain a date; this date will be added when the draft is ready and move to the _posts directory using the drafts:ready task.